+++ src/libcoretest/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2014 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -7,20 +7,25 @@
 // <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
+#![feature(globs, unsafe_destructor, macro_rules)]
 
-#![macro_escape]
-#![doc(hidden)]
+extern crate core;
+extern crate test;
+extern crate libc;
 
-macro_rules! uint_module (($T:ty, $T_SIGNED:ty, $bits:expr) => (
-
-#[unstable]
-pub static BITS : uint = $bits;
-#[unstable]
-pub static BYTES : uint = ($bits / 8);
-
-#[unstable]
-pub static MIN: $T = 0 as $T;
-#[unstable]
-pub static MAX: $T = 0 as $T - 1 as $T;
-
-))
+mod any;
+mod atomics;
+mod cell;
+mod char;
+mod cmp;
+mod finally;
+mod fmt;
+mod iter;
+mod mem;
+mod num;
+mod ops;
+mod option;
+mod ptr;
+mod raw;
+mod result;
+mod tuple;
