+++ src/libcoretest/num/i16.rs
@@ -1,4 +1,4 @@
-// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2014 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,19 +8,4 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![macro_escape]
-#![doc(hidden)]
-
-macro_rules! uint_module (($T:ty, $T_SIGNED:ty, $bits:expr) => (
-
-#[unstable]
-pub static BITS : uint = $bits;
-#[unstable]
-pub static BYTES : uint = ($bits / 8);
-
-#[unstable]
-pub static MIN: $T = 0 as $T;
-#[unstable]
-pub static MAX: $T = 0 as $T - 1 as $T;
-
-))
+int_module!(i16, i16)
